==> Code/04_discharge_conversion_level1to2.R <==
##############################################################################################
#' @title Convert units and re-calculate discharge

#' @author 
#' Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr

#' @description This function GETs data from the NEON REST Service after retrieving a 
#' transition object that must be set as an environment variable called "TRANSID" using 
#' two other environment variables that store OS (called osdataserviceurl, required) and 
#' IS (called cdsurl, optional) REST URLs.

#' @param stageData A dataframe containing stage and discharge data for a site/water year [dataframe]
#' @param dischargeData A dataframe containing distance and velocity point measurements for the 
#' site/water year that can be used to re-calculate discharge for the stageData [dataframe]

#' @return This function returns an XML file parsed into a data frame

#' @references
#' License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007

#' @export

# changelog and author contributions / copyrights
#   Kaelin M. Cawley (2017-12-07)
#     original creation
##############################################################################################
#####Discharge #######################################################################################################################

#get discharge data for focus 10 focus sites (+3 replacement sites), taken manually 
discharge_grab_subsetFINAL <- loadByProduct(dpID="DP1.20048.001", 
                                            site=c( 'KING','MAYF','COMO',"HOPB","WALK", "MART", "BLUE", "ARIK", "TOMB", "CUPE","GUIL","POSE","BLWA"),
                                            startdate="2017-01", 
                                            enddate="2019-12", 
                                            package="expanded", #basic will just give you concentrations, #expanded will give you flags 
                                            check.size = F)  ### check the size of the file before you download it 

for(i in 1:length(discharge_grab_subsetFINAL)) {assign(names(discharge_grab_subsetFINAL)[i], discharge_grab_subsetFINAL[[i]])}   #calls the table waq_instances
discharge_subsetFINAL<-as.data.frame(dsc_individualFieldData) #table that has all measurements for discharge
stage_subsetFINAL<-as.data.frame(dsc_fieldData)#table tht has stage readings from field

#These dataframes are Level 1 data, need to convert to Level 2 data using conv.calc.Q function before using in further analyses# 

#convert dataframes generated from download to names used in this function#
stageData<-stage_subsetFINAL
dischargeData<-discharge_subsetFINAL


conv.calc.Q <- function(
  stageData,
  dischargeData
){
  ##### Constants #####
  cmTom <- 1/100 #Conversion factor for centimeters to meters
  cmsToms <- 1/1000 #Conversion factor for centimetersPerSecond to metersPerSecond
  m3sTolps <- 1000 #Conversion factor for cubicMetersPerSecond to litersPerSecond
  
  #Convert all units to be the same: m, mps, and lps
  changeStageIDs <- stageData$recorduid[!stageData$streamStageUnits%in%c('m',"meter")]
  changeQIDs <- stageData$recorduid[!stageData$totalDischargeUnits%in%c('lps',"litersPerSecond")]
  changeVelocityIDs <- stageData$recorduid[!stageData$averageVelocityUnits%in%c('m/s',"meterPerSecond")]
  changeDepthIDs <- stageData$recorduid[!stageData$waterDepthUnits%in%c('m',"meter")]
  changeTapeDistIDs <- stageData$recorduid[!stageData$tapeDistanceUnits%in%c('m',"meter")]
  
  #Convert stage and discharge in the stageData table
  stageData$totalDischarge <- as.numeric(stageData$totalDischarge)
  stageData$streamStage <- as.numeric(stageData$streamStage)
  
  stageData$totalDischarge[stageData$recorduid %in% changeQIDs] <- stageData$totalDischarge[stageData$recorduid %in% changeQIDs] * m3sTolps
  stageData$streamStage[stageData$recorduid %in% changeStageIDs] <- stageData$streamStage[stageData$recorduid %in% changeStageIDs] * cmTom
  
  #Convert water depth, velocity, and tape distance
  dischargeData$tapeDistance <- as.numeric(dischargeData$tapeDistance)
  dischargeData$averageVelocity <- as.numeric(dischargeData$averageVelocity)
  dischargeData$waterDepth <- as.numeric(dischargeData$waterDepth)
  dischargeData$sectionFlow <- as.numeric(dischargeData$sectionFlow)
  dischargeData$averageVelocity <- as.numeric(dischargeData$averageVelocity)
  
  dischargeData$tapeDistance[dischargeData$recorduid %in% changeTapeDistIDs] <- dischargeData$tapeDistance[dischargeData$recorduid %in% changeTapeDistIDs] * cmTom
  dischargeData$averageVelocity[dischargeData$recorduid %in% changeVelocityIDs] <- dischargeData$averageVelocity[dischargeData$recorduid %in% changeVelocityIDs] * cmsToms
  dischargeData$waterDepth[dischargeData$recorduid %in% changeDepthIDs] <- dischargeData$waterDepth[dischargeData$recorduid %in% changeDepthIDs] * cmTom
  
  #Recalculate discharge based off of individual measurements
  allRecordID <- unique(stageData$recorduid)
  stageData$calcQ <- NA
  stageData$recalculatedL1QF <- NA
  stageData$L1DataQF <- NA
  
  for(j in seq(along = allRecordID)){
    dataForCalculations <- dischargeData[dischargeData$recorduid == allRecordID[j],]
    dataForCalculations$stationNumber <- as.numeric(dataForCalculations$stationNumber)
    
    #Check for all stations and error if not
    if(max(dataForCalculations$stationNumber) != length(dataForCalculations$stationNumber)){
      print(paste0("Warning, incomplete record for ", allRecordID[j],". Could not re-calculate discharge from measurements"))
      next
    }
    
    totalDischarge <- 0
    
    for(i in seq(along = dataForCalculations$stationNumber)){
      sectionDischarge <- NA
      depth <- dataForCalculations$waterDepth[dataForCalculations$stationNumber == i]
      velocity <- dataForCalculations$averageVelocity[dataForCalculations$stationNumber == i]
      
      previousDist <- dataForCalculations$tapeDistance[dataForCalculations$stationNumber == (i-1)]
      currentDist <- dataForCalculations$tapeDistance[dataForCalculations$stationNumber == i]
      nextDist <- dataForCalculations$tapeDistance[dataForCalculations$stationNumber == (i+1)]
      
      if(length(previousDist) < 1){
        width <- nextDist - currentDist
      }else if(length(nextDist) < 1){
        width <- currentDist - previousDist
      }else{
        width <- (nextDist - currentDist)/2 + (currentDist - previousDist)/2
      }
      #Check to see if the data line up, otherwise, re-calculate
      #Cannot recalculate all of them due to rounding issues for low flows
      roundedVelocity <- round(dataForCalculations$sectionFlow[dataForCalculations$stationNumber == i]/(width * depth * m3sTolps), 2)
      if(!is.na(roundedVelocity) && 
         roundedVelocity == dataForCalculations$averageVelocity[dataForCalculations$stationNumber == i]){
        sectionDischarge <- dataForCalculations$sectionFlow[dataForCalculations$stationNumber == i]
      }else{
        sectionDischarge <- width * depth * velocity * m3sTolps
      }
      totalDischarge <- totalDischarge + sectionDischarge
      
    }
    stageData$calcQ[stageData$recorduid == allRecordID[j]] <- totalDischarge
    
  }
  
  #Set QF for recalculated data
  stageData$recalculatedL1QF[stageData$calcQ != stageData$totalDischarge] <- 1
  stageData$recalculatedL1QF[stageData$calcQ == stageData$totalDischarge] <- 0
  stageData$L1DataQF <- stageData$dataQF
  
  return(stageData)
}

write.csv(stageData, 'Data/Discharge_Level2_QAQC.csv', row.names = FALSE)
